+++ Jump_to_Algorithm2/week5/KyoungyeonJu/17779 게리맨더링2.py
