+++ Jump_to_Algorithm2/week5/eunbiPark/11053_가장_긴_증가하는_부분수열.py
@@ -1,0 +1,11 @@
+n = int(input())
+nums = list(map(int, input().split()))
+
+dp = [1] * n
+
+for i in range(1, n):
+    for j in range(i):
+        if nums[i] > nums[j]:
+            dp[i] = max(dp[i], dp[j] + 1)
+
+print(max(dp))
